cephadm: move custom container tests to a dedicated file
